Refactor: Split FlightSearchHandler into smaller files
Refactor FlightSearchHandler.tsx into smaller, more manageable files to improve code organization and readability.
